fix(sessions): guard malformed maintenance rows
