+++ src/config/sessions/store-maintenance-total-cap.test.ts
@@ -1,4 +1,5 @@
-import { describe, expect, it } from "vitest";
+import { describe, expect, it, vi } from "vitest";
+import { applyFileBackedSessionStoreMaintenance } from "./store-maintenance-operations.js";
 import { capEntryCount, getActiveSessionMaintenanceWarning } from "./store-maintenance.js";
 import type { SessionEntry } from "./types.js";
 
@@ -10,6 +11,14 @@ function makeEntry(updatedAt: number): SessionEntry {
 
 function makeStore(entries: Array<[string, SessionEntry]>): Record<string, SessionEntry> {
   return Object.fromEntries(entries);
+}
+
+function createMaintenanceArtifacts() {
+  return {
+    archiveRemovedSessionTranscripts: async () => new Set<string>(),
+    removeRemovedSessionTrajectoryArtifacts: async () => {},
+    cleanupArchivedSessionTranscripts: async () => {},
+  };
 }
 
 describe("session maintenance total entry cap", () => {
@@ -76,5 +85,51 @@ describe("session maintenance total entry cap", () => {
         nowMs: now,
       }),
     ).toMatchObject({ wouldCap: true, wouldPrune: false });
+  });
+
+  it("uses enforcement preservation when predicting active-session eviction", async () => {
+    const now = Date.now();
+    const storePath = "/tmp/openclaw-sessions/warn-enforce-parity.json";
+    const makePressureStore = () =>
+      makeStore([
+        ["archived", { ...makeEntry(now - 2), archivedAt: now }],
+        ["active", makeEntry(now - 1)],
+        ["recent", makeEntry(now)],
+      ]);
+    const maintenanceConfig = {
+      mode: "warn" as const,
+      pruneAfterMs: 30 * DAY_MS,
+      maxEntries: 2,
+      modelRunPruneAfterMs: DAY_MS,
+      resetArchiveRetentionMs: null,
+      maxDiskBytes: null,
+      highWaterBytes: null,
+    };
+    const onWarn = vi.fn();
+
+    await applyFileBackedSessionStoreMaintenance({
+      storePath,
+      store: makePressureStore(),
+      activeSessionKey: "active",
+      maintenanceConfig,
+      onWarn,
+      log: { warn: () => {}, info: () => {} },
+      artifacts: createMaintenanceArtifacts(),
+    });
+
+    const enforcedStore = makePressureStore();
+    await applyFileBackedSessionStoreMaintenance({
+      storePath,
+      store: enforcedStore,
+      activeSessionKey: "active",
+      maintenanceConfig: { ...maintenanceConfig, mode: "enforce" },
+      log: { warn: () => {}, info: () => {} },
+      artifacts: createMaintenanceArtifacts(),
+    });
+
+    expect(onWarn).not.toHaveBeenCalled();
+    expect(enforcedStore).toHaveProperty("archived");
+    expect(enforcedStore).toHaveProperty("active");
+    expect(enforcedStore.recent).toBeUndefined();
   });
 });
